pyproject.toml
[project]
name = "diffusion-seg"
version = "0.1.0"
description = "Diffusion-based segmentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "autorootcwd>=1.0.1",
 "dvclive>=3.48.5",
 "einops>=0.8.1",
 "ipykernel==6.30.0",
 "jinja2",
 "lightning[pytorch-extra]>=2.5.5",
 "markupsafe",
 "matplotlib==3.10.3",
 "mlflow>=3.4.0",
 "monai>=1.5.0",
 "natten==0.21.0",
 "numpy==2.1.2",
 "opencv-python>=4.12.0.88",
 "pandas>=2.3.1",
 "pillow==11.0.0",
 "scikit-image>=0.25.2",
 "scipy>=1.16.1",
 "simpleitk>=2.5.2",
 "tensorboard>=2.20.0",
 "timm>=1.0.19",
 "torch>=2.7.0",
 "torchaudio>=2.7.0",
 "torchmetrics>=1.8.2",
 "torchvision>=0.22.0",
 "tqdm>=4.66.5",
 "ultralytics-thop>=2.0.17",
 "pytest>=9.0.0",
 "pytest-cov>=7.0.0",
 "torchdiffeq>=0.2.5",
 "gudhi>=3.11.0",
 "topolosses>=0.2.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/nightly/cu128"
explicit = true

[[tool.uv.index]]
url = "https://pypi.org/simple/"
default = true
